add di baseline
