+++ di_baseline/my_submission/config/gobigger_no_spatial_config.py
@@ -15,6 +15,7 @@ gobigger_dqn_config = dict(
         resize_height=160,
         resize_width=160,
         spatial=False,
+        train=True,
         manager=dict(shared_memory=False, ),
     ),
     policy=dict(
